Started work on objectfication

+++ src/main/java/org/mangorage/mangolang/compiler/LexerNode.java
@@ -6,5 +6,5 @@ import java.util.List;
 import java.util.Stack;
 
 public interface LexerNode {
-    LexerOutput handle(String[] parts, String name, Stack<BlockContext> blocks, List<Integer> out, CompilerContext ctx, InstructionSet set);
+    LexerOutput handle(String[] parts, String name, Stack<BlockContext> blocks, List<Byte> out, CompilerContext ctx, InstructionSet set);
 }
